Fix false positives, fix ctags plugin
Fixes #3789
Fixes #3792
