+++ src/Phan/Language/UnionType.php
@@ -3792,19 +3792,31 @@ class UnionType implements Serializable
             }
             // TODO: Instead of coercing string to int here, update the real type when the array is assigned to,
             // if the string is a non-literal.
-            if ($type instanceof ArrayType) {
-                foreach ($key_union_type->getTypeSet() as $key_type) {
-                    if ($key_type instanceof StringType) {
-                        // Numeric literals such as `'0'` cast to 0 when inserted as array keys.
-                        $new_real_type_builder->addType(IntType::instance(false));
-                        break;
+            //
+            // Note that array shapes with 0 elements do not have types. (tests/files/src/0461_array_key_exists.php)
+            if ($type instanceof ArrayType && $type->isPossiblyTruthy()) {
+                if ($key_union_type->isEmpty()) {
+                    $key_union_type = UnionType::fromFullyQualifiedPHPDocString('int|string');
+                } else {
+                    foreach ($key_union_type->getTypeSet() as $key_type) {
+                        if ($key_type instanceof StringType && $key_type->isPossiblyNumeric()) {
+                            // Numeric literals such as `'0'` cast to 0 when inserted as array keys.
+                            $new_real_type_builder->addType(IntType::instance(false));
+                            break;
+                        }
                     }
                 }
             }
             $new_real_type_builder->addUnionType($key_union_type);
         }
+        $type_set = $new_type_builder->getTypeSet();
+        $real_type_set = $new_real_type_builder->getTypeSet();
+        if (!$type_set && $real_type_set) {
+            $type_set = UnionType::typeSetFromString('mixed');
+        }
 
-        return UnionType::of($new_type_builder->getTypeSet(), $new_real_type_builder->getTypeSet());
+        $result = UnionType::of($type_set, $real_type_set);
+        return $result;
     }
 
     /**
