+++ src/Phan/Plugin/Internal/CtagsPlugin.php
@@ -4,22 +4,17 @@ declare(strict_types=1);
 
 namespace Phan\Plugin\Internal;
 
-use Exception;
 use Phan\CLI;
 use Phan\CodeBase;
 use Phan\Config;
 use Phan\Language\Element\Clazz;
 use Phan\Language\Element\Flags;
-use Phan\Language\FileRef;
-use Phan\Language\FQSEN;
-use Phan\Language\FQSEN\FullyQualifiedClassElement;
 use Phan\Language\FQSEN\FullyQualifiedClassName;
 use Phan\Library\FileCache;
+use Phan\Plugin\Internal\CtagsPlugin\CtagsEntry;
+use Phan\Plugin\Internal\CtagsPlugin\CtagsEntrySet;
 use Phan\PluginV3;
 use Phan\PluginV3\FinalizeProcessCapability;
-
-use function is_string;
-use function strlen;
 
 /**
  * This plugin generates a ctags file from the files parsed by Phan.
@@ -92,12 +87,15 @@ class CtagsPlugin extends PluginV3 implements
         if ($class->isPHPInternal()) {
             return;
         }
-        $entries->add(new CtagsEntry(
-            $class->getName(),
-            $class->getContext(),
-            CtagsEntry::KIND_CLASS,
-            CtagsEntry::generateScopeLabelForNamespace($class->getContext()->getNamespace())
-        ));
+        if (!$class->isAnonymous()) {
+            // Generate a ctags entry for named classes (no possible tag for `new class extends ...`)
+            $entries->add(new CtagsEntry(
+                $class->getName(),
+                $class->getContext(),
+                CtagsEntry::KIND_CLASS,
+                CtagsEntry::generateScopeLabelForNamespace($class->getContext()->getNamespace())
+            ));
+        }
         $label = CtagsEntry::generateScopeLabelForFQSEN($class_fqsen);
 
         foreach ($class->getConstantMap($code_base) as $const) {
@@ -134,10 +132,11 @@ class CtagsPlugin extends PluginV3 implements
             ));
         }
 
-        foreach ($class->getPropertyMap($code_base) as $property_fqsen => $property) {
+        foreach ($class->getPropertyMap($code_base) as $property) {
             if ($property->isDynamicProperty()) {
                 continue;
             }
+            $property_fqsen = $property->getFQSEN();
             if ($property_fqsen !== $property->getRealDefiningFQSEN()) {
                 continue;
             }
@@ -206,154 +205,6 @@ EOT;
             $contents .= "$entry\n";
         }
         return $contents;
-    }
-}
-
-/**
- * Represents a sorted set of unique ctags entries for definitions of elements
- */
-class CtagsEntrySet
-{
-    /** @var array<string, CtagsEntry> */
-    private $entries = [];
-    /**
-     * @return array<string, CtagsEntry> a sorted map of sorted entries.
-     * This sorts the entries when it is called.
-     */
-    public function toArray(): array
-    {
-        \uksort($this->entries, 'strcmp');
-        return $this->entries;
-    }
-
-    /**
-     * Record an occurrence of an element definition.
-     */
-    public function add(CtagsEntry $entry): void
-    {
-        if (!$entry->isValid()) {
-            return;
-        }
-        $lookup = $entry->name . "\0" . $entry;
-        $this->entries[$lookup] = $entry;
-    }
-}
-
-/**
- * Represents a single ctags entry.
- *
- * @phan-read-only
- */
-class CtagsEntry
-{
-    public const KIND_CLASS = 'c';
-    /** "define" in ctags is also used for constants */
-    public const KIND_CONSTANT = 'd';
-    public const KIND_FUNCTION = 'f';
-    public const KIND_PROPERTY = 'p';
-
-    /** @var string the name of the tag */
-    public $name;
-    /** @var FileRef the file and line the tag referred to */
-    public $context;
-    /** @var string the kind of tag*/
-    public $kind;
-    /** @var ?string the scope, e.g. `function:some_function`, `namespace:MyNS\SubNS`, `class:SomeClass`, etc. */
-    public $scope;
-    /** @var string the code fragment to locate the tag */
-    public $fragment;
-    /** @var bool is this entry valid */
-    private $is_valid = false;
-
-    public function __construct(string $name, FileRef $context, string $kind, ?string $scope, string $fragment = null)
-    {
-        if ($context->isPHPInternal()) {
-            return;
-        }
-        if (!\is_string($fragment)) {
-            $path = Config::projectPath($context->getFile());
-            try {
-                $entry = FileCache::getOrReadEntry($path);
-            } catch (Exception $e) {
-                CLI::printToStderr("Failed to read $path: {$e->getMessage()}\n");
-                return;
-            }
-
-            $fragment = $entry->getLines()[$context->getLineNumberStart()] ?? '';
-        }
-        if (\strlen($fragment) === 0) {
-            CLI::printToStderr("Empty line for generating ctags for name=$name at $context kind=$kind\n");
-            return;
-        }
-        $this->name = $name;
-        $this->context = $context;
-        $this->kind = $kind;
-        $this->scope = $scope;
-        $this->fragment = \rtrim($fragment);
-        $this->is_valid = true;
-    }
-
-    /**
-     * Is this ctags entry valid?
-     */
-    public function isValid(): bool
-    {
-        return $this->is_valid;
-    }
-
-    public function __toString()
-    {
-        /*
-         * > The name of the "kind:" field can be omitted.  This is to reduce the size of
-         * > the tags file by about 15%.  A program reading the tags file can recognize the
-         * > "kind:" field by the missing ':'.
-         */
-        $ctags_path = Config::projectPath($this->context->getFile());
-        $common = \sprintf(
-            "%s\t%s\t%s\t%s\tline:%d",
-            $this->name,
-            $ctags_path,
-            self::escapeFragment($this->fragment),
-            $this->kind,
-            $this->context->getLineNumberStart()
-        );
-        if (is_string($this->scope) && strlen($this->scope) > 0) {
-            $common .= "\t$this->scope";
-        }
-        return $common;
-    }
-
-    /**
-     * Escape a code fragment for the ctags file format.
-     */
-    public static function escapeFragment(string $fragment): string {
-        $escaped = \str_replace(['\\', '/', "\0"], ['\\\\', '\/', '\\0'], $fragment);
-        return '/^' . $escaped . '$/;"';
-    }
-
-    /**
-     * Generate a scope label for the provided FQSEN
-     */
-    public static function generateScopeLabelForFQSEN(?FQSEN $fqsen): ?string
-    {
-        if ($fqsen instanceof FullyQualifiedClassName) {
-            return 'class:' . \ltrim((string)$fqsen, '\\');
-        } elseif ($fqsen instanceof FullyQualifiedClassElement) {
-            return self::generateScopeLabelForFQSEN($fqsen->getFullyQualifiedClassName());
-        }
-        return null;
-    }
-
-    /**
-     * Generate a scope label for the provided namespace.
-     */
-    public static function generateScopeLabelForNamespace(string $namespace): ?string
-    {
-        $namespace = \ltrim($namespace, "\\");
-        if (strlen($namespace) > 0)  {
-            return "namespace:$namespace";
-        }
-        return null;
     }
 }
