+++ tests/files/expected/0786_non_empty_array_phpdoc.php.expected
@@ -4,5 +4,5 @@
 %s:31 PhanTypeMismatchArgument Argument 1 ($values) is [] of type array{} but \accept_non_empty_array() takes non-empty-array<int,string> defined at %s:24
 %s:35 PhanTypeMismatchArgument Argument 1 ($values) is ['first'=>'x'] of type array{first:'x'} but \accept_non_empty_array() takes non-empty-array<int,string> defined at %s:24
 %s:42 PhanTypeMismatchArgumentInternal Argument 1 ($numerator) is $v of type \stdClass|string but \intdiv() takes int
-%s:43 PhanTypeMismatchArgumentInternal Argument 1 ($obj) is $k of type int|string but \spl_object_id() takes object
+%s:43 PhanTypeMismatchArgumentInternalProbablyReal Argument 1 ($obj) is $k of type int|string but \spl_object_id() takes object
 %s:48 PhanTypeMismatchArgument Argument 1 ($values) is [] of type array{} but \accept_non_empty_array2() takes non-empty-array<mixed,\stdClass>|non-empty-array<mixed,string> defined at %s:39
