+++ tests/files/expected/0865_array_key_int_or_string.php.expected
@@ -1,0 +1,1 @@
+%s:6 PhanImpossibleTypeComparisonInLoop Impossible attempt to check if $a of type array is identical to $k of type int|string in a loop body (likely a false positive)
